Make the GET DIAGNOSTICS counter 64 bit safe
Along the way update all places which work with the
number of tuples, one way or another.

+++ src/include/cdb/cdbdispatchresult.h
@@ -107,7 +107,7 @@ typedef struct CdbDispatchResult
 	bool wasCanceled;
 
 	/* num rows rejected in SREH mode */
-	int	numrowsrejected;
+	int64	numrowsrejected;
 } CdbDispatchResult;
 
 /*
